+++ tests/test_historical_client.py
@@ -170,7 +170,7 @@ class TestHistoricalClient:
         assert call["params"] == [
             ("dataset", "GLBX.MDP3"),
             ("start", "2020-12-28T13:00:00+00:00"),
-            ("end", "2020-12-29T13:00:00+00:00"),
+            ("end", "2020-12-29T13:01:00+00:00"),
             ("symbols", "ESH1"),
             ("schema", "definition"),
             ("stype_in", "native"),